+++ packages/core/src/use-dynamic-animation/index.ts
@@ -46,10 +46,7 @@ export default function useDynamicAnimation(
     activeStyle.current.value = initialState()
   }
 
-  const __state = useSharedValue(
-    activeStyle.current.value,
-    false // don't rebuild it (for older versions)
-  )
+  const __state = useSharedValue(activeStyle.current.value)
 
   const controller = useRef<UseDynamicAnimationState>()
 
@@ -61,24 +58,13 @@ export default function useDynamicAnimation(
       },
       animateTo(nextStateOrFunction) {
         'worklet'
-        // const runAnimation = (nextStyleObject: DynamicStyleProp) => {
-        //   if (nextStyleObject) {
-        //     // activeStyle.current.value = nextStyleObject
-        //     __state.value = nextStyleObject
-        //   }
-        // }
+
         const nextStyle =
           typeof nextStateOrFunction === 'function'
             ? nextStateOrFunction(__state.value)
             : nextStateOrFunction
 
         __state.value = nextStyle
-        // if (typeof nextStateOrFunction === 'function') {
-        //   // similar to setState, let people compose a function that takes in the current value and returns the next one
-        //   runAnimation(nextStateOrFunction(this.current as DynamicStyleProp))
-        // } else {
-        //   runAnimation(nextStateOrFunction)
-        // }
       },
     }
   }