+++ packages/skeleton/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0](https://github.com/nandorojo/moti/compare/v0.10.1...v0.11.0) (2021-05-17)
+
+**Note:** Version bump only for package @motify/skeleton
+
+
+
+
+
+# [0.10.0](https://github.com/nandorojo/moti/compare/v0.9.0...v0.10.0) (2021-04-21)
+
+**Note:** Version bump only for package @motify/skeleton
+
+
+
+
+
+# [0.9.0](https://github.com/nandorojo/moti/compare/v0.8.2...v0.9.0) (2021-04-21)
+
+**Note:** Version bump only for package @motify/skeleton
+
+
+
+
+
 ## [0.8.2](https://github.com/nandorojo/moti/compare/v0.8.1...v0.8.2) (2021-04-02)
 
 **Note:** Version bump only for package @motify/skeleton